requirements.txt
Flask==1.1.2
Pillow==9.0.1
requests==2.27.1
gunicorn==20.1.0
boto3==1.20.54